refactor(big-picture): streamline library hook and page by consolidating update logic into the hook
